backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
ultralytics
torch
torchvision
scikit-learn
numpy
Pillow
scipy
